+++ CHANGELOG.md
@@ -38,6 +38,9 @@ NOTE: Add new changes BELOW THIS COMMENT.
 
 - The `edge` channel has been switched to the new UI and versioning scheme.
 
+- The dashboard no longer shows zero-valued statistics for disabled Safe
+  Browsing, Parental Control, and Safe Search features ([#689]).
+
 ### Deprecated
 
 - `strict_sni_check` is now deprecated.
@@ -51,6 +54,7 @@ NOTE: Add new changes BELOW THIS COMMENT.
 [go-1.26.6]: https://groups.google.com/g/golang-announce/c/94pEornpRlI
 [#7514]: https://github.com/AdguardTeam/AdGuardHome/issues/7514
 [#8183]: https://github.com/AdguardTeam/AdGuardHome/issues/8183
+[#689]:  https://github.com/AdguardTeam/AdGuardHome/issues/689
 
 <!--
 NOTE: Add new changes ABOVE THIS COMMENT.